pyproject.toml
[tool.poetry]
name = "fructus-temporum-task"
version = "0.1.0"
description = "fructus temporum test task"
authors = ["Сергей Гриценко <gritsenko.serge.2013@yandex.ru>"]
license = "MIT"
readme = "README.md"
packages = [{include = "fructus_temporum_task"}]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.4"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.7"
djoser = "^2.2.0"
drf-spectacular = "^0.26.4"
djangorestframework = "^3.14.0"
beautifulsoup4 = "^4.12.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
autopep8 = "2.0.0"
flake8 = "5.0.4"
flake8-blind-except = "0.2.1"
flake8-broken-line = "0.6.0"
flake8-bugbear = "22.9.23"
flake8-builtins = "2.0.0"
flake8-class-attributes-order = "0.1.3"
flake8-cognitive-complexity = "0.1.0"
flake8-commas = "2.1.0"
flake8-comprehensions = "3.10.0"
flake8-debugger = "4.1.2"
flake8-eradicate = "1.4.0"
flake8-functions = "0.0.7"
flake8-mutable = "1.2.0"
flake8-print = "5.0.0"
flake8-pyproject = "^1.2.2"
flake8-pytest = "1.4"
flake8-pytest-style = "1.6.0"
flake8-quotes = "3.3.1"
flake8-string-format = "0.3.0"
flake8-variables-names = "0.0.5"
isort = "^5.11.4"
mypy = "^1.0.0"
pre-commit = "3.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
skip-string-normalization=true
skip-magic-trailing-comma=false
line-length=89

[tool.flake8]
ignore=['W503', 'CCE001']
exclude=['.git', 'env/', 'venv/', 'temp/', '*.md', '__pycache__/']
max-cognitive-complexity=10
max-line-length=89
class_attributes_order=['field', 'meta_class', 'nested_class', 'magic_method', 'property_method', 'static_method', 'class_method', 'method', 'private_method']

[tool.isort]
known_first_party=['users', 'api', 'core', 'bookmarks']
profile='black'

[tool.mypy]
exclude=['venv/']
ignore_missing_imports=true
